requirements.txt
autopep8==1.4.4
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2
django-elasticsearch-dsl==7.1.1
django-import-export==2.0
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
tablib==1.0.0
urllib3==1.25.7
